% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{bootstrap_predict_common}
\alias{bootstrap_predict_common}
\title{Internal function}
\usage{
bootstrap_predict_common(object, newdata, type = c("mean"), extrap = TRUE, ...)
}
\description{
predicting bootstrapGradientForest and combinedBootstrapGF
have a lot of overlap.
}
\details{
This function pulls together common operations.
}
