% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{combine_gfbootstrap_dist}
\alias{combine_gfbootstrap_dist}
\title{Combined gf_dist}
\usage{
combine_gfbootstrap_dist(gf_combine, x_samples = 100)
}
\arguments{
\item{gf_combine}{combinedBootstrapGF object}

\item{x_samples}{number of points along each cumimp curve for
deciding the best offsets}
}
\value{
data.frame, bootstraps by predictors
}
\description{
Calculate the distance between bootstrapped
samples of combined gradientForest models.
}
