% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{gfbootstrap_dist}
\alias{gfbootstrap_dist}
\title{gfbootstrap distance matrix}
\usage{
gfbootstrap_dist(gf_boot, x_samples = 100)
}
\arguments{
\item{gf_boot}{a bootstrapGradientForest object}

\item{x_samples}{integer number of points sampled along each cumulative importance curve, evenly spaced.
higher values will give more accurate answers, but consume more time and resources}
}
\value{
Returns a data.frame of areas between curves for each pair and each predictor.
}
\description{
Given a list of gradient forest objects,
gfbootstrap_dist calculates the area bewteen curves for
each pair of gf objects.
}
\details{
gfbootstrap_dist() assumes that every GF model has the same
set of predictor variables.

If an offsets vector is provided, it is used to calculate the result.
}
