% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{gfbootstrap_offsets}
\alias{gfbootstrap_offsets}
\title{Calculate opimal offsets}
\usage{
gfbootstrap_offsets(x)
}
\description{
The distance values returned by \code{gfbootstrap_dist()}
are used to calculate the optimal offsets.
Given a predictor, an offset is calculated for each
gf cumulative importance curve,
such that the total area between the curves is
minimised.
}
\details{
\param x the data.frame returned by gfbootstrap_dist()
}
