% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{gg_bootstrapGF}
\alias{gg_bootstrapGF}
\title{Predict bootstrapped GF object}
\usage{
gg_bootstrapGF(
  x,
  vars = names(importance(x$gf_list[[1]], type = "Weighted", sorted = TRUE)),
  n_curves = 10,
  debug = TRUE
)
}
\description{
Given a fitted gfbootstrap object, predict new points.
}
\details{
Unlike most predict functions, you get multiple output rows per input row.
The id_cols parameter allows you to specify which columns should be used as identifiers for each input row.
The total number of output rows is gfbootstrap$nbootstraps * nrow(predict_points)

plot cumimp for bootStrapGradientForest

Returns a ggplot object, ready for plotting
}
