% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{hubert_gamma}
\alias{hubert_gamma}
\title{Huberts Gamma statistic}
\usage{
hubert_gamma(sim_mat, member_mat, norm_z = TRUE)
}
\arguments{
\item{sim_mat}{a square matrix equivalent, all values in the range 0 to 1}

\item{member_mat}{square matrix equivalent, 0 if a pair of sites are in different
clusters, 1 if the pair are in the same cluster.}

\item{norm_z}{whether to normalize the sim_mat and member_mat into z-scores. See Tseng and Kao 2003}
}
\value{
the Hubert Gamma score
}
\description{
Calculates Huberts Gamma statistic
}
\details{
norm_z rescales both x and y
by subtracting the mean and dividing by the
standard deviation.
The norm_z variant is from Tseng and Kao 2003.

The original Huberts Gamma statistic (Hubert and Levin 1976)
uses x and y as is.

Zhao et al. 2006 proposes another variant, where
y is defined as the distance between the cluster
centers of site i and site j.
This variant is harder to calculate and gives
a knee rather than a peak.
This variant is not implemented.

This version requires a full matrix for both sim_mat and
member_mat. Using other forms is much more complicated.
}
