% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{importance.bootstrapGradientForest}
\alias{importance.bootstrapGradientForest}
\title{Get variable importance from a bootstrap Gradient Forest}
\usage{
\method{importance}{bootstrapGradientForest}(
  x,
  type = c("Accuracy", "Impurity", "Weighted", "Raw", "Species")[3],
  sort = TRUE
)
}
\arguments{
\item{x}{A bootstrapGradientForest object}

\item{type}{What kind of importance to use. See importance.gradientForest in the gradientForest package for more details.}

\item{sort}{Should the predictors be sorted in order of importance?}
}
\description{
Finds the overall importance of each predictor by averaging
the importance of each predictor from each bootstrap run.
Predictors with more importance are better predictors of the
response variables.
}
\details{
returns a named vector of predictor importances, possibly sorted.
}
