% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{importance.combinedBootstrapGF}
\alias{importance.combinedBootstrapGF}
\title{Get variable importance from a combined bootstrapped Gradient Forest}
\usage{
\method{importance}{combinedBootstrapGF}(x, type = c("Weighted", "Raw", "Species")[1], sort = TRUE)
}
\arguments{
\item{x}{A combinedBootstrapGF object}

\item{type}{What kind of importance to use. See importance.combinedGradientForest in the gradientForest package for more details.}

\item{sort}{Should the predictors be sorted in order of importance?}
}
\description{
Finds the overall importance of each predictor by averaging
the importance of each predictor from each bootstrap run.
Predictors with more importance are better predictors of the
response variables.
}
\details{
returns a named vector of predictor importances, possibly sorted.
}
