% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{importance_bootstrap_common}
\alias{importance_bootstrap_common}
\title{Internal function}
\usage{
importance_bootstrap_common(x, type, sort)
}
\description{
calculating importance for bootstrapGradientForest and
combinedBootstrapGF has a lot of overlap.
}
\details{
This function pulls together common operations.
}
