% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{combinedBootstrapGF}
\alias{combinedBootstrapGF}
\title{Combine bootstrapped gradient forest objects}
\usage{
combinedBootstrapGF(
  ...,
  n_samp,
  x_samples = 100,
  nbin = 101,
  method = 2,
  standardize = c("before", "after")[1],
  weight = c("uniform", "species", "rsq.total", "rsq.mean", "site", "site.species",
    "site.rsq.total", "site.rsq.mean")[3]
)
}
\arguments{
\item{n_samp}{number of bootstrap samples}

\item{x_samples}{number of points along each cumimp curve for
deciding the best offsets}

\item{nbin}{number of bins for the cumimp curves}

\item{method}{see gradientForest::combinedGradientForest}

\item{standardize}{see gradientForest::combinedGradientForest}

\item{weight}{see gradientForest::combinedGradientForest}
}
\value{
combinedBootstrapGF object
}
\description{
As for gradientForest::combinedGradientForest(),
takes two or more bootstrapGradientForest objects
and combines them.
}
\details{
Each bootStrapGradientForest object has many individual
GF objects, and one GF model from each bootstrapGradientForest
parameter is taken and used to create one sample in the
combined bootstrapped Gradient Forest model.

Each bootstrap sample in the combined bootstrapped Gradient Forest
model is a combinedGradientForest object, with one GF model from
each of the individual bootStrapGradientForest objects.

The key difference to combinedGradientForest
is that weights must be specified in this call,
not at the predict.combinedGradientForest() call,
so offsets for each curve can be calculated now.
}
