% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{gg_bootstrapGF}
\alias{gg_bootstrapGF}
\title{plot cumimp for bootStrapGradientForest}
\usage{
gg_bootstrapGF(
  x,
  vars = names(importance(x$gf_list[[1]], type = "Weighted", sorted = TRUE)),
  n_curves = 10,
  debug = TRUE
)
}
\arguments{
\item{x}{a bootStrapGradientForest object}

\item{vars}{names of predictor variables to use}

\item{n_curves}{Number of curves to plot. If less than the number of bootstrap samples, then samples are chosen randomly}

\item{debug}{additional plots and messages if TRUE}
}
\value{
Returns a ggplot object, ready for plotting.
}
\description{
plot cumimp for bootStrapGradientForest
}
