% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{predict.bootstrapGradientForest}
\alias{predict.bootstrapGradientForest}
\title{Predict bootstrapGradientForest}
\usage{
\method{predict}{bootstrapGradientForest}(object, newdata, type = c("mean"), extrap = TRUE, ...)
}
\arguments{
\item{object}{a bootstrapGradientForest object}

\item{newdata}{data.frame of new observations to predict. If NULL, use observations from object.}

\item{type}{is a vector containing elements from
c("mean", "variance", "points", "weight")

"mean" gives the unweighted mean of the points (default, for
consistency with other predict functions)

"variance" gives the unweighted variance of the points

"points" returns all the points, so for each new x row,
k (number of bootstrapped models) rows will be returned

"weight" gives the R^2 performance of each bootstrapped model.

Due to the fact that many rows of output can be returned
for each row of input, x_row is included in the returned
data.frame to match inputs to outputs.}

\item{extrap}{passed on to predict.gradientForest
possible valuse are NA, TRUE, FALSE, or a number in the range \link{0,1}}

\item{...}{arguments passed to gradientForest::cumimp()
TODO: This code is rather spaghetti, eg. activity 4 depends on 2 and 3,
then activity 5 depends on 1 and 4}
}
\value{
A long form data.frame, with $type giving the prediction type.
}
\description{
Generic S3 function for predicting bootstrapGradientForest
objects.
}
\details{
The behaviour is very similar to predict.gradientForest,
but due to the way bootstrapGradientForest is intended
to be used, an extra parameter, \code{type} is added.
}
