% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfbootstrap.R
\name{predict.combinedBootstrapGF}
\alias{predict.combinedBootstrapGF}
\title{Predict combinedBootstrapGF}
\usage{
\method{predict}{combinedBootstrapGF}(object, newdata, type = c("mean"), extrap = TRUE, ...)
}
\arguments{
\item{object}{a combinedBootstrapGF object}

\item{newdata}{data.frame of new observations to predict. If NULL, use observations
from the first GF object.}

\item{type}{is a vector containing elements from
c("mean", "variance", "points", "weight")

"mean" gives the unweighted mean of the points (default, for
consistency with other predict functions)

"variance" gives the unweighted variance of the points

"points" returns all the points, so for each new x row,
k (number of bootstrapped models) rows will be returned

"weight" gives the R^2 performance of each bootstrapped model.

Due to the fact that many rows of output can be returned
for each row of input, x_row is included in the returned
data.frame to match inputs to outputs.}

\item{extrap}{passed on to predict.gradientForest
possible valuse are NA, TRUE, FALSE, or a number in the range \link{0,1}}

\item{...}{arguments passed to gradientForest::cumimp()}
}
\value{
A long form data.frame, with $type giving the prediction type.
}
\description{
Generic S3 function for predicting combinedBootstrapGF
objects.
}
\details{
The behaviour is very similar to predict.gradientForest,
but due to the way bootstrapGradientForest is intended
to be used, an extra parameter, \code{type} is added.
}
